+++ api/app/connectors/monero/__init__.py
@@ -1,3 +1,0 @@
-from .connector import MoneroConnector
-
-__all__ = ["MoneroConnector"]